requirements.txt
django==5.2.*
django-environ
django-redis
psycopg2-binary
djangorestframework
djangorestframework-simplejwt
dj-rest-auth
django-cors-headers
django-filter
django-cleanup
whitenoise
gunicorn
